Fix potential MissingBackpressureException
In case notifications/indications come too quickly to be pushed through the bridge to non-native side there will be emitted a `MissingBackpressureException` from RxJava. To mitigate the problem a buffering backpressure strategy is added which should help in most common situations.

Pros:
- No notifications will be discarded
Cons:
- Potential `OutOfMemoryException` in case of prolonged, high speed notifications

Technically backpressure strategy should be exposed on the API level but can be introduced at a later moment.