Fixed BluetoothGattCharacteristic UUID collision
It may be that a device contain two or more `BluetoothGattCharacteristic` objects that share the same `UUID`. In this case wrong `BluetoothGattCharacteristic` could been chosen if the `UUID` has been passed to `RxAndroidBle`. Fixed by using the `BluetoothGattCharacteristic` object as a parameter and relaxing accesibility of it in `Characteristic` class from `private` to `package protected`.